requirements.txt
box-box>=0.0.1
jsonschema
json-schema-to-class>=0.0.8
torch
torch-utils
